server/requirements.txt
fastapi
redis
celery
uvicorn
numpy
pandas
scikit-learn
transformers
torch
kafka-python
python-dotenv
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
certifi
python-docx
huggingface_hub
requests
beautifulsoup4
youtube-transcript-api
PyMuPDF
pydantic